Feat: UserDetailService Implements to Custom Class
