--- src/main/java/com/knj/mirou/base/security/CustomOAuth2User.java
@@ -1,25 +1,0 @@
-package com.knj.mirou.base.security;
-
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-
-import java.util.Collection;
-import java.util.Map;
-
-public class CustomOAuth2User extends User implements OAuth2User {
-
-    public CustomOAuth2User(String username, String password, Collection<? extends GrantedAuthority> authorities) {
-        super(username, password, authorities);
-    }
-
-    @Override
-    public Map<String, Object> getAttributes() {
-        return null;
-    }
-
-    @Override
-    public String getName() {
-        return getUsername();
-    }
-}
