+++ src/main/java/com/knj/mirou/base/security/CustomUserDetailsService.java
@@ -1,0 +1,27 @@
+package com.knj.mirou.base.security;
+
+import com.knj.mirou.boundedContext.member.model.entity.Member;
+import com.knj.mirou.boundedContext.member.repository.MemberRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = false)
+public class CustomUserDetailsService implements UserDetailsService {
+
+    private final MemberRepository memberRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        Member member = memberRepository.findByLoginId(username).orElseThrow(()
+                -> new UsernameNotFoundException("%s는 찾을 수 없는 회원명 입니다.".formatted(username)));
+
+        return new User(member.getLoginId(), "1234", member.getGrantedAuthorities());
+    }
+}
